requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.15.0
cryptography==36.0.1
cycler==0.10.0
Django==3.2.3
django-rest-knox==4.2.0
djangorestframework==3.13.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
psycopg2==2.9.2
pycparser==2.21
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
python-decouple==3.5
pytz==2021.1
pyzmq==22.1.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
wincertstore==0.2
